Fix to filter to only suppliers who have made a successful application
Previously the script filtered for all suppliers who had an interest in a
framework and who hadn't signed their framework agreement. This
includes the users we wanted to email, but mistakenly also sent emails
to supplier users who had started an application but not completed it,
or those who were unsuccessful in their application.

We want to email just the suppliers who have made a successful
application and not yet signed their agreement. This script is run
towards the end of standstill so by that point CCS will have made
decisions on all discretionary applications and all suppliers who
have successfully applied will have `onFramework` set to `True`
in the database.

To get to the users we want we can get all the suppliers who haven't
signed their agreement in the same way, but crucially we now also
filter for just the suppliers who are marked as on the framework.

https://trello.com/c/AWM8eJYY/2159-p4-incident-ccsrequests-fwd-you-must-sign-your-digital-outcomes-and-specialists-5-framework-award-form