Add notify-suppliers-with-signed-framework-agreement.py
We want to be able to send an email to all suppliers who have signed
their framework agreement when a framework goes live. This commit adds a
script that sends a given Notify email template to all suppliers with a
signed framework agreement.

It was almost entirely copied from
`scripts/framework-applications/remind-suppliers-to-sign-framework-agreement.py`,
and then modified to have the personalisations needed for Notify template
"50b Services are live (DOS 5)" (98a9754d-9310-436a-9e7a-cc7504ecd92d).