Create script to get the ID of G12 recovery drafts
As part of the G12 recovery, we need a list of the IDs of the recovery drafts. This will allow us to hide irrelevant drafts to avoid confusion.

We have a CSV containing the unnormalised service names. For each, get all drafts for that supplier that have a matching service name when normalised. Output the same CSV with the new data in a new column.

Further processing will be needed to turn this into a 1:1 mapping.