Choose between multiple G12 recovery drafts
Some G12 recovery have multiple corresponding drafts. This can be one for each time we ran the CSV import scripts (up to 5), in addition to the original draft service created by the supplier.

We need to choose one of them for each. So choose the one with the most completed answers. We think this should be the most useful to suppliers.

I've tested this manually, and it seems to work correctly.