Add awscli to Docker image using pipx
We want to install awscli into a separate virtual environment so that
the version of botocore it requires does not conflict with other
libraries.

The easiest way to do this is to use [pipx], which creates a new virtual
environment, installs awscli into it, and then links the `aws`
executable into `~/.local/bin`.

[pipx]: https://pipxproject.github.io/pipx/