Add script to evaluate G12 recovery supplier declaration
We have one supplier in the G12 recovery process who has not yet
signed the framework agreement. In order for them to do this, we
need to evaluate their answers to the declaration questions and
mark them as being on G-Cloud 12 if they pass.

This script borrows heavily from
digitalmarketplace-scripts/blob/master/scripts/framework-applications/mark-definite-framework-results.py
to do so for this single supplier.

Note: Their declaration *looks* fine from a quick manual comparison
against the schema. Therefore I've not written code to export their
declaration answers if they don't pass automatically. If that's
needed, we'll want to borrow from
scripts/framework-applications/export-framework-results-reasons.py

https://trello.com/c/QwO6dH3A/794-1-make-sure-recovery-supplier-not-yet-on-framework-can-sign-framework-agreement