Copy uncopied G12 service documents
During the G12 recovery, we published 17 services without copying their service documents to the documents bucket. So the document links did not work.

Find the affected documents, copy them to the correct bucket, and update the affected services with the new URL.

I've done a dry-run of this, and it looked as if it should do the right thing.