Add hash of command line args to email_engine reference
Based on a suggestion by @gidsg. Fixes unexpected behaviour where
running the same script with different flags has the same reference and
logfile (and therefore might not re-run because the previous run
succeeded, or maybe send the previous notifications ignoring the new
arguments). Adds a hash of the command line arguments to the reference,
so two different runs will have a slightly different reference, but runs
with the same script and same arguments will have the same reference.