Pass notification hash to Notify
We want to send a unique reference to Notify for every email we send.
The reference should be reproducible to allow us to re-create and look
up an email's status in our logs at a future date. This is useful when
investigating 2nd line requests, or in the face of a potential legal
challenge from a supplier.

Scripts not using the email engine already construct this unique reference.

Add a property `EmailNotification.sha256_hash` which computes the
SHA256 hash on the string representation of the object.

`EmailNotification` already implements the `__hash__()` method,
however this is salted with a random value which changes between Python
processes[[1]] and therefore cannot be used to recreate a reference
at a future date.

[1]:https://docs.python.org/3/reference/datamodel.html#object.__hash__