Get buyers for research who've used a lot
The admin frontend will give us a list of all buyers who've opted in to being contacted about user research. However, for this particular bit of research, we only want to contact those who've used the Digital Specialists lot.

Just a one-off script for now. If we end up having to run this again, we should consider moving the functionality into admin frontend.