Create a script to make Dependabot PRs easier
Currently, we need to review, approve and merge upwards some tens of Dependabot PRs each week. The actual reviewing is not very hard - its usually obvious whether to approve a Dependabot PR.

The main problem is that the process involves so many manual clicks. To go from the search page to the PR, approve, and the merge takes at least 4 clicks accross 3 pages.

Thus, I'm creating this script to automate the parts of this process that can be automated. It searches for Dependabot PRs, skipping those that are obviously not ready for merging. It then opens the PR in the browser for a human to do the review. The human then just needs to enter 'y' for the PR to be approved and merged.

I think this will cut down quite significantly on the time and effort of dealing with Dependabot PRs.