Add initial script to generate data for the DMp
The first stage of my firebreak project.

We've discovered that the API will start against a local database as
long as the required tables exist - it doesn't matter if there's no
data in any of the tables. This means we can use the API client to
insert our randomised data and benefit from all the validation and
error handling that the API brings.

Currently this script only adds a single buyer user account, but it
sets up a proof of concept for future steps.