Show unsuccessful checks
When reviewing Dependabot PRs, it's handy to know more about the PRs where the check status is not successful. If it's in progress, I don't need to do anything. But I want to know if the check completed but failed, because that means I need to intervene.

Show the details of checks that weren't successful. The format isn't very pretty, but should be good enough for a developer to work out what's going on.