Script getting emails from mailchimp for 'coming'
This is currently a manual process. For DOS 5, we made a mistake during the manual process. We exported the wrong list, so didn't notify everyone we ought to.

Create a script that we can use for this process in future. This should reduce the risk of similar incidents in future.

A human will still need to merge this list with the list of supplier email addresses. However, I think that's much less risky, so not a priority for scripting.