Get stats on direct award usage
We've had a request for the number of buyers who started the procurement journey in the past year. This script should answer this question for G-Cloud. We already have this data for DOS in the public export of opportunity data.

I'm also including direct award projects where no outcome was reported. This is because it is likely that some of those had an outcome that was not reported to us.