Email stats snapshot
We've not currently got a way to upload stats to Google Drive and recreating it looks to be slow. So instead enhance the snapshot script to it can send the stats in an email. It sends it to a Google Group in the CCS domain so that people who want can access them.

Notify template: https://www.notifications.service.gov.uk/services/a9333c2c-e3a5-4eb3-bb28-86f3920c1900/templates/493ede76-d9c8-40b8-b543-1127ccb60674