Get bot PRs from each repo individually
Recently, I've found that the script has not picked up Dependabot PRs in some repos. I suspect this is because either the number of PRs or the number of repos is too large. Work around this by running the search indivually for each repo, to guarantee we get all the PRs. It makes the script a bit slower to run, but more comprehensive.