Upload service edit report to S3
We've currently got a manual process to create the report and email it to a particular category admin once a month. I'm starting to get bored of this, so would like to automate it. I've added an endpoint so that admins can download the report, so this report alters the script to upload it to S3.